Worked on Cart order  value total and cart item delete features

+++ FoodiesLounge/Controllers/MenuItemsController.cs
@@ -28,15 +28,19 @@ namespace FoodiesLounge.Controllers
         public IActionResult Delete(int id)
         {
             var result = _db.GetFirstOrDefault(u=> u.Id == id);
-            string webRoothPath = _webHost.WebRootPath;
-            var OLdImagePath = Path.Combine(webRoothPath, result.Image.TrimStart('\\'));
-            if (System.IO.File.Exists(OLdImagePath))
+            if (result.Image != null)
             {
-                System.IO.File.Delete(OLdImagePath);
+                string webRoothPath = _webHost.WebRootPath;
+                var OLdImagePath = Path.Combine(webRoothPath, result.Image.TrimStart('\\'));
+                if (System.IO.File.Exists(OLdImagePath))
+                {
+                    System.IO.File.Delete(OLdImagePath);
+                }
+                _db.Remove(result);
+                _db.Save();
+                return Json(new { success = true, message = " Delete successful" });
             }
-            _db.Remove(result);
-            _db.Save();
-            return Json(new { success = true, message = " Delete successful" });
+            return Json(new { success = true, message = " File not found" });
         }
     }
 }
